remove uninteractiblePiece, size positionable-piece correctly
